-> Add : V1 Nearby Search with retrofit for support Api 19. -> Need to fix takePicture (actually crash on api 19)

+++ app/src/androidTest/java/com/openclassrooms/realestatemanager/ContentProviderTest.kt
@@ -79,7 +79,7 @@ class ContentProviderTest {
         id = realEstateDatabase.RealEstateDao().insert(realEstate = RealEstate(
             type = "Appartement",
             price = 1000999,
-            surface = 500.toFloat(),
+            surface = 500,
             nbRooms = 12,
             nbBathrooms = 2,
             nbBedrooms = 3,
